Split API modules, add focus-trap, refactor UI
Extract frontend API surface from lib/api/today into focused modules (analytics, integrationClients, medications, search, settings, templates, mealPlans, shoppingLists) using fetchWithAuth/parseJsonResponse and zod validations. Add useFocusTrap hook and wire it into multiple modals/overlays. Refactor SettingsPage into small sections/components: IntegrationClientsSection, OAuthSessionsSection, UserPreferencesSection. Improve backend background-loop logging boundaries to avoid stopping periodic tasks. Update tests and query client imports to match the new module layout.
