+++ frontend/src/lib/api/today.ts
@@ -11,27 +11,26 @@ export {
   sleep,
   withAuthHeader,
 } from "@/lib/api/http";
+import {
+  medicationHistoryItemSchema,
+  medicationMutationResponseSchema,
+  medicationTodayItemSchema,
+  type MedicationDoseStatus,
+  type MedicationHistoryItem,
+  type MedicationMutationResponse,
+  type MedicationTodayItem,
+} from "@/lib/api/medications";
 import { z } from "zod";
 
 export type TaskStatus = "pending" | "in_progress" | "completed" | "skipped";
 export type ChoreStatus = "pending" | "completed" | "skipped";
-export type MedicationDoseStatus = "scheduled" | "taken" | "skipped" | "missed";
 export type PlannedItemModuleKey =
   | "shopping_list"
   | "meal_planning"
   | "recurring_grocery"
   | "shared_calendar";
 
-export interface MedicationTodayItem {
-  medication_dose_instance_id: number;
-  medication_plan_id: number;
-  name: string;
-  instructions: string;
-  scheduled_at: string;
-  status: MedicationDoseStatus;
-}
-
-export type MedicationHistoryItem = MedicationTodayItem;
+export type { MedicationDoseStatus, MedicationHistoryItem, MedicationMutationResponse, MedicationTodayItem };
 
 export interface RoutineTodayItem {
   task_instance_id: number;
@@ -173,19 +172,6 @@ const plannedItemModuleKeySchema = z.enum([
 
 const taskStatusSchema = z.enum(["pending", "in_progress", "completed", "skipped"]);
 const choreStatusSchema = z.enum(["pending", "completed", "skipped"]);
-const medicationDoseStatusSchema = z.enum(["scheduled", "taken", "skipped", "missed"]);
-
-const medicationTodayItemSchema = z.object({
-  medication_dose_instance_id: z.number(),
-  medication_plan_id: z.number(),
-  name: z.string(),
-  instructions: z.string(),
-  scheduled_at: z.string(),
-  status: medicationDoseStatusSchema,
-});
-
-const medicationHistoryItemSchema = medicationTodayItemSchema;
-
 const routineTodayItemSchema = z.object({
   task_instance_id: z.number(),
   routine_template_id: z.number(),
@@ -315,26 +301,6 @@ const taskMutationResponseSchema = z.object({
   completed_at: z.string().nullable(),
 });
 
-const medicationMutationResponseSchema = z.object({
-  medication_dose_instance_id: z.number(),
-  status: medicationDoseStatusSchema,
-  scheduled_date: z.string(),
-});
-
-const medicationPlanSchema = z.object({
-  id: z.number(),
-  name: z.string(),
-  instructions: z.string(),
-  start_date: z.string(),
-  schedule_time: z.string(),
-  every_n_days: z.number(),
-  is_active: z.boolean(),
-});
-
-const medicationHistoryResponseSchema = z.object({
-  history: z.array(medicationHistoryItemSchema),
-});
-
 export interface ChoreMutationResponse {
   chore_instance_id: number;
   status: ChoreStatus;
@@ -350,130 +316,6 @@ export interface TaskMutationResponse {
   due_at: string | null;
   completed_at: string | null;
 }
-
-export interface MedicationMutationResponse {
-  medication_dose_instance_id: number;
-  status: MedicationDoseStatus;
-  scheduled_date: string;
-}
-
-export interface MedicationPlan {
-  id: number;
-  name: string;
-  instructions: string;
-  start_date: string;
-  schedule_time: string;
-  every_n_days: number;
-  is_active: boolean;
-}
-
-export interface MedicationPlanInput {
-  name: string;
-  instructions: string;
-  start_date: string;
-  schedule_time: string;
-  every_n_days: number;
-}
-
-export interface MedicationPlanUpdateInput extends MedicationPlanInput {
-  is_active: boolean;
-}
-
-export interface MedicationHistoryResponse {
-  history: MedicationHistoryItem[];
-}
-
-export interface IntegrationClient {
-  id: number;
-  name: string;
-  rate_limit_per_minute: number;
-  is_active: boolean;
-}
-
-export interface IntegrationClientInput {
-  name: string;
-  rate_limit_per_minute: number;
-}
-
-export interface IntegrationClientCreateResponse extends IntegrationClient {
-  api_key: string;
-  client_id: string;
-  client_secret: string;
-  token_url: string;
-}
-
-export interface RoutineTemplate {
-  id: number;
-  name: string;
-  description: string | null;
-  start_date: string;
-  every_n_days: number;
-  due_time: string | null;
-  is_active: boolean;
-  created_at: string;
-}
-
-export interface RoutineTemplateInput {
-  name: string;
-  description?: string | null;
-  start_date: string;
-  every_n_days: number;
-  due_time?: string | null;
-  is_active: boolean;
-}
-
-export interface ChoreTemplate {
-  id: number;
-  name: string;
-  description: string | null;
-  start_date: string;
-  every_n_days: number;
-  is_active: boolean;
-  created_at: string;
-}
-
-export interface ChoreTemplateInput {
-  name: string;
-  description?: string | null;
-  start_date: string;
-  every_n_days: number;
-  is_active: boolean;
-}
-
-const integrationClientSchema = z.object({
-  id: z.number(),
-  name: z.string(),
-  rate_limit_per_minute: z.number(),
-  is_active: z.boolean(),
-});
-
-const integrationClientCreateResponseSchema = integrationClientSchema.extend({
-  api_key: z.string(),
-  client_id: z.string(),
-  client_secret: z.string(),
-  token_url: z.string(),
-});
-
-const routineTemplateSchema = z.object({
-  id: z.number(),
-  name: z.string(),
-  description: z.string().nullable(),
-  start_date: z.string(),
-  every_n_days: z.number(),
-  due_time: z.string().nullable(),
-  is_active: z.boolean(),
-  created_at: z.string(),
-});
-
-const choreTemplateSchema = z.object({
-  id: z.number(),
-  name: z.string(),
-  description: z.string().nullable(),
-  start_date: z.string(),
-  every_n_days: z.number(),
-  is_active: z.boolean(),
-  created_at: z.string(),
-});
 
 export async function fetchToday(signal?: AbortSignal): Promise<TodayPayload> {
   return getJson("/api/today", TodayResponseSchema, signal, 1, "Unable to load today's data");
@@ -620,500 +462,6 @@ export async function skipRoutineTask(taskInstanceId: number): Promise<TaskMutat
     headers: { Accept: "application/json" },
   });
   return parseJsonResponse(response, "Request failed", false, taskMutationResponseSchema);
-}
-
-export async function listMedicationPlans(signal?: AbortSignal): Promise<MedicationPlan[]> {
-  return getJson("/api/medications", z.array(medicationPlanSchema), signal);
-}
-
-export async function createMedicationPlan(input: MedicationPlanInput): Promise<MedicationPlan> {
-  return sendJson("POST", "/api/medications", input, medicationPlanSchema);
-}
-
-export async function updateMedicationPlan(
-  medicationPlanId: number,
-  input: MedicationPlanUpdateInput,
-): Promise<MedicationPlan> {
-  return sendJson("PUT", `/api/medications/${medicationPlanId}`, input, medicationPlanSchema);
-}
-
-export async function deleteMedicationPlan(medicationPlanId: number): Promise<void> {
-  const response = await fetchWithAuth(`/api/medications/${medicationPlanId}`, {
-    method: "DELETE",
-    headers: { Accept: "application/json" },
-  });
-
-  if (!response.ok) {
-    // Error-only parse; successful deletes return 204 with no JSON body.
-    await parseJsonResponse<never>(response, "Request failed", false);
-  }
-}
-
-export async function fetchMedicationHistory(
-  signal?: AbortSignal,
-): Promise<MedicationHistoryItem[]> {
-  const payload = await getJson("/api/medication-doses/history", medicationHistoryResponseSchema, signal);
-  return payload.history;
-}
-
-export interface CalendarFeedResponse {
-  token: string;
-  feed_url: string;
-}
-
-const calendarFeedResponseSchema = z.object({
-  token: z.string(),
-  feed_url: z.string(),
-});
-
-export async function fetchCalendarFeed(signal?: AbortSignal): Promise<CalendarFeedResponse> {
-  const response = await fetchWithAuth("/api/calendar/feed", {
-    headers: { Accept: "application/json" },
-    signal,
-  });
-  return parseJsonResponse(response, "Failed to load calendar feed", true, calendarFeedResponseSchema);
-}
-
-export async function regenerateCalendarFeed(): Promise<CalendarFeedResponse> {
-  const response = await fetchWithAuth("/api/calendar/feed/regenerate", {
-    method: "POST",
-    headers: { Accept: "application/json" },
-  });
-  return parseJsonResponse(response, "Failed to regenerate calendar feed", false, calendarFeedResponseSchema);
-}
-
-export async function listIntegrationClients(signal?: AbortSignal): Promise<IntegrationClient[]> {
-  const response = await fetchWithAuth("/api/integrations/clients", {
-    headers: { Accept: "application/json" },
-    signal,
-  });
-  return parseJsonResponse(response, "Request failed", true, z.array(integrationClientSchema));
-}
-
-export async function createIntegrationClient(
-  input: IntegrationClientInput,
-): Promise<IntegrationClientCreateResponse> {
-  const response = await fetchWithAuth("/api/integrations/clients", {
-    method: "POST",
-    headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(input),
-  });
-  return parseJsonResponse(response, "Request failed", false, integrationClientCreateResponseSchema);
-}
-
-export async function rotateIntegrationClient(
-  clientId: number,
-): Promise<IntegrationClientCreateResponse> {
-  const response = await fetchWithAuth(`/api/integrations/clients/${clientId}/rotate`, {
-    method: "POST",
-    headers: { Accept: "application/json" },
-  });
-  return parseJsonResponse(
-    response,
-    "Failed to rotate integration client",
-    false,
-    integrationClientCreateResponseSchema,
-  );
-}
-
-export async function revokeIntegrationClient(clientId: number): Promise<void> {
-  const response = await fetchWithAuth(`/api/integrations/clients/${clientId}`, {
-    method: "DELETE",
-  });
-  if (!response.ok) {
-    // Error-only parse; successful revocations return 204 with no JSON body.
-    await parseJsonResponse<never>(response, "Failed to revoke integration client");
-  }
-}
-
-export async function listRoutineTemplates(signal?: AbortSignal): Promise<RoutineTemplate[]> {
-  const response = await fetchWithAuth("/api/templates/routines", {
-    headers: { Accept: "application/json" },
-    signal,
-  });
-  return parseJsonResponse(response, "Request failed", true, z.array(routineTemplateSchema));
-}
-
-export async function createRoutineTemplate(input: RoutineTemplateInput): Promise<RoutineTemplate> {
-  const response = await fetchWithAuth("/api/templates/routines", {
-    method: "POST",
-    headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(input),
-  });
-  return parseJsonResponse(response, "Request failed", false, routineTemplateSchema);
-}
-
-export async function updateRoutineTemplate(
-  routineTemplateId: number,
-  input: RoutineTemplateInput,
-): Promise<RoutineTemplate> {
-  const response = await fetchWithAuth(`/api/templates/routines/${routineTemplateId}`, {
-    method: "PUT",
-    headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(input),
-  });
-  return parseJsonResponse(response, "Request failed", false, routineTemplateSchema);
-}
-
-export async function deleteRoutineTemplate(routineTemplateId: number): Promise<void> {
-  const response = await fetchWithAuth(`/api/templates/routines/${routineTemplateId}`, {
-    method: "DELETE",
-    headers: { Accept: "application/json" },
-  });
-
-  if (!response.ok) {
-    // Error-only parse; successful deletes return 204 with no JSON body.
-    await parseJsonResponse<never>(response, "Request failed", false);
-  }
-}
-
-export async function listChoreTemplates(signal?: AbortSignal): Promise<ChoreTemplate[]> {
-  const response = await fetchWithAuth("/api/templates/chores", {
-    headers: { Accept: "application/json" },
-    signal,
-  });
-  return parseJsonResponse(response, "Request failed", true, z.array(choreTemplateSchema));
-}
-
-export async function createChoreTemplate(input: ChoreTemplateInput): Promise<ChoreTemplate> {
-  const response = await fetchWithAuth("/api/templates/chores", {
-    method: "POST",
-    headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(input),
-  });
-  return parseJsonResponse(response, "Request failed", false, choreTemplateSchema);
-}
-
-export async function updateChoreTemplate(
-  choreTemplateId: number,
-  input: ChoreTemplateInput,
-): Promise<ChoreTemplate> {
-  const response = await fetchWithAuth(`/api/templates/chores/${choreTemplateId}`, {
-    method: "PUT",
-    headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(input),
-  });
-  return parseJsonResponse(response, "Request failed", false, choreTemplateSchema);
-}
-
-export async function deleteChoreTemplate(choreTemplateId: number): Promise<void> {
-  const response = await fetchWithAuth(`/api/templates/chores/${choreTemplateId}`, {
-    method: "DELETE",
-    headers: { Accept: "application/json" },
-  });
-
-  if (!response.ok) {
-    // Error-only parse; successful deletes return 204 with no JSON body.
-    await parseJsonResponse<never>(response, "Request failed", false);
-  }
-}
-
-// --- User Settings ---
-
-export interface UserSettings {
-  timezone: string;
-  default_snooze_days: number;
-  medication_reminder_minutes: number;
-  quiet_hours_start: string | null;
-  quiet_hours_end: string | null;
-  push_overdue_chores_enabled: boolean;
-  push_medication_reminders_enabled: boolean;
-  push_missed_medications_enabled: boolean;
-}
-
-export interface UserSettingsPatch {
-  timezone?: string;
-  default_snooze_days?: number;
-  medication_reminder_minutes?: number;
-  quiet_hours_start?: string | null;
-  quiet_hours_end?: string | null;
-  push_overdue_chores_enabled?: boolean;
-  push_medication_reminders_enabled?: boolean;
-  push_missed_medications_enabled?: boolean;
-}
-
-const userSettingsSchema = z.object({
-  timezone: z.string(),
-  default_snooze_days: z.number(),
-  medication_reminder_minutes: z.number(),
-  quiet_hours_start: z.string().nullable(),
-  quiet_hours_end: z.string().nullable(),
-  push_overdue_chores_enabled: z.boolean(),
-  push_medication_reminders_enabled: z.boolean(),
-  push_missed_medications_enabled: z.boolean(),
-});
-
-export async function fetchUserSettings(signal?: AbortSignal): Promise<UserSettings> {
-  const response = await fetchWithAuth("/api/users/me/settings", {
-    headers: { Accept: "application/json" },
-    signal,
-  });
-  return parseJsonResponse(response, "Request failed", true, userSettingsSchema);
-}
-
-export async function updateUserSettings(patch: UserSettingsPatch): Promise<UserSettings> {
-  const response = await fetchWithAuth("/api/users/me/settings", {
-    method: "PATCH",
-    headers: { Accept: "application/json", "Content-Type": "application/json" },
-    body: JSON.stringify(patch),
-  });
-  return parseJsonResponse(response, "Failed to update settings", false, userSettingsSchema);
-}
-
-// --- Analytics ---
-
-export type AnalyticsPeriod = "week" | "month" | "quarter" | "year";
-
-export interface DailyCount {
-  date: string;
-  completed: number;
-  total: number;
-  completion_rate: number;
-}
-
-export interface ChoreStreak {
-  chore_id: number;
-  name: string;
-  current_streak: number;
-  longest_streak: number;
-}
-
-export interface SkippedChore {
-  chore_id: number;
-  name: string;
-  skip_count: number;
-}
-
-export interface ChoreStats {
-  completion_rate: number;
-  total_completed: number;
-  total_scheduled: number;
-  daily_completions: DailyCount[];
-  streaks: ChoreStreak[];
-  most_skipped: SkippedChore[];
-}
-
-export interface DailyAdherence {
-  date: string;
-  taken: number;
-  total: number;
-  adherence_rate: number;
-}
-
-export interface MedicationAnalyticsStats {
-  adherence_rate: number;
-  total_taken: number;
-  total_scheduled: number;
-  daily_adherence: DailyAdherence[];
-}
-
-export interface PlannedItemAnalyticsStats {
-  completion_rate: number;
-  total_completed: number;
-  total_scheduled: number;
-  daily_completions: DailyCount[];
-}
-
-export interface RoutineStreak {
-  routine_id: number;
-  name: string;
-  current_streak: number;
-  longest_streak: number;
-}
-
-export interface RoutineAnalyticsStats {
-  completion_rate: number;
-  total_completed: number;
-  total_scheduled: number;
-  daily_completions: DailyCount[];
-  streaks: RoutineStreak[];
-}
-
-export interface AnalyticsSummary {
-  period: AnalyticsPeriod;
-  start_date: string;
-  end_date: string;
-  chores: ChoreStats;
-  medications: MedicationAnalyticsStats;
-  planned_items: PlannedItemAnalyticsStats;
-  routines: RoutineAnalyticsStats;
-}
-
-const dailyCountSchema = z.object({
-  date: z.string(),
-  completed: z.number(),
-  total: z.number(),
-  completion_rate: z.number(),
-});
-
-const analyticsSummarySchema = z.object({
-  period: z.enum(["week", "month", "quarter", "year"]),
-  start_date: z.string(),
-  end_date: z.string(),
-  chores: z.object({
-    completion_rate: z.number(),
-    total_completed: z.number(),
-    total_scheduled: z.number(),
-    daily_completions: z.array(dailyCountSchema),
-    streaks: z.array(z.object({
-      chore_id: z.number(),
-      name: z.string(),
-      current_streak: z.number(),
-      longest_streak: z.number(),
-    })),
-    most_skipped: z.array(z.object({
-      chore_id: z.number(),
-      name: z.string(),
-      skip_count: z.number(),
-    })),
-  }),
-  medications: z.object({
-    adherence_rate: z.number(),
-    total_taken: z.number(),
-    total_scheduled: z.number(),
-    daily_adherence: z.array(z.object({
-      date: z.string(),
-      taken: z.number(),
-      total: z.number(),
-      adherence_rate: z.number(),
-    })),
-  }),
-  planned_items: z.object({
-    completion_rate: z.number(),
-    total_completed: z.number(),
-    total_scheduled: z.number(),
-    daily_completions: z.array(dailyCountSchema),
-  }),
-  routines: z.object({
-    completion_rate: z.number(),
-    total_completed: z.number(),
-    total_scheduled: z.number(),
-    daily_completions: z.array(dailyCountSchema),
-    streaks: z.array(z.object({
-      routine_id: z.number(),
-      name: z.string(),
-      current_streak: z.number(),
-      longest_streak: z.number(),
-    })),
-  }),
-});
-
-export async function fetchAnalyticsSummary(
-  period: AnalyticsPeriod = "week",
-  signal?: AbortSignal,
-): Promise<AnalyticsSummary> {
-  const response = await fetchWithAuth(`/api/analytics/summary?period=${period}`, {
-    headers: { Accept: "application/json" },
-    signal,
-  });
-  return parseJsonResponse(response, "Failed to load analytics", true, analyticsSummarySchema);
-}
-
-// --- Search ---
-
-export interface RoutineSearchResult {
-  id: number;
-  name: string;
-  description: string | null;
-  is_active: boolean;
-  created_at: string;
-}
-
-export interface ChoreSearchResult {
-  id: number;
-  name: string;
-  description: string | null;
-  priority: string;
-  tags: string[];
-  is_active: boolean;
-  created_at: string;
-}
-
-export interface MedicationSearchResult {
-  id: number;
-  name: string;
-  instructions: string;
-  is_active: boolean;
-  created_at: string;
-}
-
-export interface PlannedItemSearchResult {
-  id: number;
-  title: string;
-  notes: string | null;
-  planned_for: string;
-  priority: string;
-  tags: string[];
-  is_done: boolean;
-  created_at: string;
-}
-
-export interface SearchResponse {
-  query: string;
-  routine_templates: RoutineSearchResult[];
-  chore_templates: ChoreSearchResult[];
-  medication_plans: MedicationSearchResult[];
-  planned_items: PlannedItemSearchResult[];
-}
-
-const searchResponseSchema = z.object({
-  query: z.string(),
-  routine_templates: z.array(z.object({
-    id: z.number(),
-    name: z.string(),
-    description: z.string().nullable(),
-    is_active: z.boolean(),
-    created_at: z.string(),
-  })),
-  chore_templates: z.array(z.object({
-    id: z.number(),
-    name: z.string(),
-    description: z.string().nullable(),
-    priority: z.string(),
-    tags: z.array(z.string()),
-    is_active: z.boolean(),
-    created_at: z.string(),
-  })),
-  medication_plans: z.array(z.object({
-    id: z.number(),
-    name: z.string(),
-    instructions: z.string(),
-    is_active: z.boolean(),
-    created_at: z.string(),
-  })),
-  planned_items: z.array(z.object({
-    id: z.number(),
-    title: z.string(),
-    notes: z.string().nullable(),
-    planned_for: z.string(),
-    priority: z.string(),
-    tags: z.array(z.string()),
-    is_done: z.boolean(),
-    created_at: z.string(),
-  })),
-});
-
-export async function searchItems(query: string, signal?: AbortSignal): Promise<SearchResponse> {
-  const response = await fetchWithAuth(`/api/search?q=${encodeURIComponent(query)}`, {
-    headers: { Accept: "application/json" },
-    signal,
-  });
-  return parseJsonResponse(response, "Search failed", true, searchResponseSchema);
 }
 
 // --- Planned item reschedule ---
