+++ frontend/tests/features/settings/SettingsPage.test.tsx
@@ -22,12 +22,14 @@ const pwaMock = vi.hoisted(() => ({
   subscribeInstallPrompt: vi.fn(),
 }));
 
-vi.mock("@/lib/api/today", () => ({
+vi.mock("@/lib/api/integrationClients", () => ({
   createIntegrationClient: apiMock.createIntegrationClient,
-  isRetryableApiError: () => false,
   listIntegrationClients: apiMock.listIntegrationClients,
   revokeIntegrationClient: apiMock.revokeIntegrationClient,
   rotateIntegrationClient: apiMock.rotateIntegrationClient,
+}));
+
+vi.mock("@/lib/api/settings", () => ({
   fetchUserSettings: apiMock.fetchUserSettings,
   updateUserSettings: apiMock.updateUserSettings,
 }));
